+++ src/c++/test/test_breakpoint_genotyper.cpp
@@ -27,6 +27,7 @@
 #include "genotyping/BreakpointGenotyper.hh"
 #include "genotyping/GenotypingParameters.hh"
 #include "gmock/gmock.h"
+#include <math.h>
 #include <string>
 #include <vector>
 
@@ -36,37 +37,45 @@ using namespace genotyping;
 
 TEST(BreakpointGenotyper, ThrowsWhenWrongNumberOfReadcounts)
 {
-    const double read_depth = 40.0;
-    const int32_t read_length = 100;
+    double read_depth = 40.0;
+    int32_t read_length = 100;
+    double depth_sd = sqrt(read_depth * 5);
     const vector<string> alleles = { "REF", "ALT" };
 
     auto param = std::unique_ptr<GenotypingParameters>(new GenotypingParameters(alleles, 2));
     BreakpointGenotyper genotyper(param);
-    ASSERT_ANY_THROW(genotyper.genotype(read_depth, read_length, {}));
-    ASSERT_ANY_THROW(genotyper.genotype(read_depth, read_length, { 10 }));
+    const BreakpointGenotyperParameter b_param(read_depth, read_length, depth_sd, false);
+    ASSERT_ANY_THROW(genotyper.genotype(b_param, {}));
+    ASSERT_ANY_THROW(genotyper.genotype(b_param, { 10 }));
 }
 
 TEST(BreakpointGenotyper, GenotypesWellCoveredBreakpoints)
 {
-    const double read_depth = 40.0;
-    const int32_t read_length = 100;
     const vector<string> alleles = { "REF", "ALT" };
     auto param = std::unique_ptr<GenotypingParameters>(new GenotypingParameters(alleles, 2));
     BreakpointGenotyper genotyper(param);
 
-    EXPECT_EQ("0/0", (string)genotyper.genotype(read_depth, read_length, { 20, 0 }));
-    EXPECT_EQ("0/1", (string)genotyper.genotype(read_depth, read_length, { 20, 20 }));
-    EXPECT_EQ("1/1", (string)genotyper.genotype(read_depth, read_length, { 0, 20 }));
+    double read_depth = 40.0;
+    int32_t read_length = 100;
+    double depth_sd = 20;
+    const BreakpointGenotyperParameter b_param(read_depth, read_length, depth_sd, false);
+
+    EXPECT_EQ("0/0", (string)genotyper.genotype(b_param, { 20, 0 }));
+    EXPECT_EQ("0/1", (string)genotyper.genotype(b_param, { 20, 20 }));
+    EXPECT_EQ("1/1", (string)genotyper.genotype(b_param, { 0, 20 }));
 
     // test chrX
     auto haploid_param = std::unique_ptr<GenotypingParameters>(new GenotypingParameters(alleles, 1));
     BreakpointGenotyper haploid_genotyper(haploid_param);
-    EXPECT_EQ("1", (string)haploid_genotyper.genotype(read_depth, read_length, { 0, 20 }));
+    EXPECT_EQ("1", (string)haploid_genotyper.genotype(b_param, { 0, 20 }));
 
-    EXPECT_FLOAT_EQ(0.0080560343, (float)genotyper.genotype(read_depth, read_length, { 0, 20 }).coverage_test_pvalue);
+    EXPECT_FLOAT_EQ(0.24825223, (float)genotyper.genotype(b_param, { 0, 20 }).coverage_test_pvalue);
+
+    const BreakpointGenotyperParameter b_poisson_param(read_depth, read_length, depth_sd, true);
+    EXPECT_FLOAT_EQ(0.0080560343, (float)genotyper.genotype(b_poisson_param, { 0, 20 }).coverage_test_pvalue);
 
     const vector<string> alleles2 = { "REF", "ALT1", "ALT2", "ALT3", "ALT4" };
     auto param2 = std::unique_ptr<GenotypingParameters>(new GenotypingParameters(alleles2, 2));
     BreakpointGenotyper genotyper_q(param2);
-    EXPECT_EQ("1/3", (string)genotyper_q.genotype(read_depth, read_length, { 1, 20, 2, 20, 2 }));
+    EXPECT_EQ("1/3", (string)genotyper_q.genotype(b_param, { 1, 20, 2, 20, 2 }));
 }